+++ main.py
@@ -2,8 +2,8 @@ from litestar import Litestar
 from litestar.openapi import OpenAPIConfig
 from litestar_granian import GranianPlugin
 
-from src.routes import route_handlers
 from src.core.config import settings
+from src.routes import route_handlers
 
 
 app = Litestar(
@@ -14,7 +14,7 @@ app = Litestar(
         settings.sqlalchemy.plugin,
     ],
     openapi_config=OpenAPIConfig(
-        version="1.0.0",
+        version="0.1.0",
         title="Тестовое задание",
         summary="REST API для управления пользователями",
     ),
