+++ src/core/exceptions.py
@@ -3,12 +3,13 @@ from litestar.exceptions import HTTPException
 
 
 class UserNotFound(HTTPException):
-    def __init__(self, user_id: int):
-        """Custom UserNotFound Exception
+    """Custom UserNotFound Exception
 
-        Parameters
-            user_id: int - Идентификатор пользователя
-        """
+    Parameters
+        user_id: int - Идентификатор пользователя
+    """
+
+    def __init__(self, user_id: int):
         super().__init__(
             detail=f"No user found by id={user_id}",
             status_code=status_codes.HTTP_404_NOT_FOUND,
